디테일 페이지 리팩토링
